Refactored Token filter to check id and email
