+++ Backend/src/main/java/com/SYSC4806_Group13/SYSC4806_Project/Model/Repositories/UserRepository.java
@@ -13,4 +13,6 @@ public interface UserRepository extends CrudRepository<User, Long> {
     Optional<User> findByEmail(String email);
 
     User findUserById(long id);
+
+    Optional<User> findByIdAndEmail(long id, String email);
 }
