+++ Backend/src/main/java/com/SYSC4806_Group13/SYSC4806_Project/Service/UserService.java
@@ -8,4 +8,8 @@ public interface UserService {
     UserDetails loadUserByUsername(String email) throws UsernameNotFoundException;
 
     UserDetails loadUserById(Long id);
+
+    UserDetails loadUserByIdAndEmail(long id, String email);
+
+
 }
