+++ Backend/src/main/java/com/SYSC4806_Group13/SYSC4806_Project/Service/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.SYSC4806_Group13.SYSC4806_Project.Service;
 
 
-import com.SYSC4806_Group13.SYSC4806_Project.Exception.ResourceNotFoundException;
+import com.SYSC4806_Group13.SYSC4806_Project.Exceptions.ResourceNotFoundException;
 import com.SYSC4806_Group13.SYSC4806_Project.Model.DataModel.User;
 import com.SYSC4806_Group13.SYSC4806_Project.Model.Repositories.UserRepository;
 import com.SYSC4806_Group13.SYSC4806_Project.Security.UserPrincipal;
@@ -10,8 +10,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Slf4j
 @Service
@@ -33,6 +31,15 @@ public class UserServiceImpl implements UserService {
     public UserDetails loadUserById(Long id) {
         User user = userRepository.findById(id).orElseThrow(
                 () -> new ResourceNotFoundException("User", "id", id)
+        );
+
+        return UserPrincipal.create(user);
+    }
+
+
+    public UserDetails loadUserByIdAndEmail(long id, String email) {
+        User user = userRepository.findByIdAndEmail(id, email).orElseThrow(
+                () -> new ResourceNotFoundException("User", "id email", id+" "+email)
         );
 
         return UserPrincipal.create(user);
