--- Backend/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.security.oauth2.client.registration.google.client-id=1
-spring.security.oauth2.client.registration.google.client-secret=1
-spring.security.oauth2.client.registration.google.redirectUri=1
-spring.security.oauth2.client.registration.google.scope=1
-app.tokenSecret=1
-app.tokenExpirationMsec=1
-app.authorizedRedirectUris=1
